Create regression test based on a golden master file
This adds a regression characterization test, which consists in
running the new code and compare its output with the original one,
called golden master file.

This way, we can reduce writing the spec and refactor faster. However,
the specs should be added down the line, and also, generating this
golden master file require some effort to include all the required
cases.